requirements.txt
streamlit
instructor
google-genai
google-generativeai
pypdf
pydantic
jsonref
pandas
plotly